requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
psycopg2-binary==2.9.10
isort==6.0.1
django-environ==0.12.0